docs: 优化README和gitignore配置，移除运行时二进制文件

+++ CosyVoice/cosyvoice/llm/llm.py
@@ -226,7 +226,7 @@ class TransformerLM(torch.nn.Module):
 class Qwen2Encoder(torch.nn.Module):
     def __init__(self, pretrain_path):
         super().__init__()
-        self.model = Qwen2ForCausalLM.from_pretrained(pretrain_path)
+        self.model = Qwen2ForCausalLM.from_pretrained(pretrain_path, torch_dtype=torch.float32)
 
     def forward(self, xs: torch.Tensor, xs_lens: torch.Tensor):
         T = xs.size(1)
